refactor: update file name to meteor
