+++ packages/themes/src/index.ts
@@ -1,7 +1,7 @@
 // codegen:start { preset: barrel, include: ./*.ts }
 export * from './classic.js'
-export * from './halloween.js'
 export * from './legacy.js'
+export * from './meteor.js'
 export * from './mui.js'
 export * from './pcte.js'
 // codegen:end
